backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python-headless
requests
python-dotenv
